Issue-30: Actions of initial state are not executed
Motivation
----------
Sometimes, users want to be able to execute the entry actions of the
state machine when this one starts. This is the implementation of that
feature to avoid devs rewriting always the same workaround.

Modifications
-------------
New flag in StateMachineConfig to decide whether or not to execute the
entry actions of the state machine when it starts. StateMachine
constructor updated to check that flag and call the enter method when it
is true. Two unit tests to verify the default behavior is backward
compatible and the new behavior is working.
Version upgraded from 2.5.1-SNAPSHOT to 2.5.2-SNAPSHOT

Result
------
One can choose to execute entry actions of the initial state of the
state machine when this one starts.