Issue-15: Support action on transition (1/3)
Motivation
----------

UML state machine describes a transition as:

                ---- event [ guard ] / action --->

Current implementation of stateless4j only supports actions on entry and
exit which is not the same.

Modifications
-------------

This commit adds code so one can attach an action to each 'normal'
transition.  'Normal' means that I duplicated the two methods
'StateConfiguration.permit' and 'StateConfiguration.permitIf' and gave
the copy an additional parameter of type 'Action'. This action is then
executed (when the transition is taken) after the onExit-method of the
current state and before the onEntry-method of the next state.

Result
------

One can add an action to a static transition between two distinct nodes.
This action gets called when the transition is taken.