Issue-15: Support action on transition (2/3)
Motivation
----------

UML state machine describes a transition as:

                ---- event [ guard ] / action --->

Current implementation of stateless4j only supports actions on entry and
exit which is not the same.

Modifications
-------------

This commit adds the same functionality as the last but this time for
re-entrant transitions. In detail this means that I duplicated the two
methods 'StateConfiguration.permitReentry' and
'StateConfiguration.permitReentryIf' and gave the copy an additional
parameter of type 'Action'. This action is then executed (when the
transition is taken) between the onExit-method and the onEntry-method of
the re-entered state.

Result
------

One can add an action to a re-entrant transition (returning to the exact
same node). This action gets called when the transition is taken.