Issue-15: Support action on transition (3/3)
Motivation
----------

UML state machine describes a transition as:

                ---- event [ guard ] / action --->

Current implementation of stateless4j only supports actions on entry and
exit which is not the same.

Modifications
-------------

This commit adds actions to the dynamic transitions. In detail this means
that I duplicated all the methods 'StateConfiguration.permitDynamic' and
'StateConfiguration.permitDynamicIf' (all in all that make eight new
methods).
I gave the copies additional parameters of type 'Action',
'Action1', 'Action2' and 'Action3' (depending on the method). The given
action is then executed (when the transition is taken) between the
onExit-method and the onEntry-method of the re-entered state. It will receive
the parameters of the trigger (similar to the destinationStateSelectors).

Result
------

One can add an action to a dynamic transition (those who compute their target
based on the trigger's parameter). This action gets called when the transition
is taken. The action will receive the parameters of the trigger.