Issue-17: Support internal transition
Motivation
----------

Sometimes you need to execute an action if a trigger is fired but you
don't want to change the state. This is done by adding an internal
transition (which were not yet implemented).
With this commit I implemented the possibility to add guarded and
unguarded internal transitions.

Modifications
-------------

This commit adds the two methods `StateConfiguration.permitInternal` and
`StateConfiguration.permitInternalIf`. Furthermore I've extended
`TriggerBehaviour` to contain a method `isInternal`.
Based on the result of this method, the state machine performs a
transition or only executes the action.
There are also unit tests and I dropped a short line in the README.

Result
------

One can add guarded and unguarded internal transitions which execute an
action when taken but don't change the current state.