Issue-17 allowed some cleanups
Now where there are internal transitions there is no need for a special
IgnoredTriggerBehaviour: An 'ignored transition' is nothing else but an
internal transition without an action.

Another simplification is the refactoring from
`TriggerBehaviour.resultsInTransitionFrom` towards
`TriggerBehaviour.transitionsTo` (the information if a transition is
performed or not is now taken from `TriggerBehaviour.isInternal`). This
way we can completely drop `OutVar` which was only used to handle 'the
second return value'.